chore: reduce bundle size
